+++ D4Opaque.h
@@ -59,11 +59,11 @@ public:
 
     virtual void clear_local_data();
 
-    virtual unsigned int width(bool = false) const { return sizeof(vector<uint8_t>); }
+    virtual uint64_t width(bool = false) const { return sizeof(vector<uint8_t>); }
 
     // Return the length of the stored data or zero if no string has been
     // stored in the instance's internal buffer.
-    virtual int length() const { return d_buf.size(); }
+    virtual uint64_t length() const { return (uint64_t)d_buf.size(); }
 
     // DAP2
     virtual bool serialize(ConstraintEvaluator &, DDS &, Marshaller &, bool = true) {
@@ -81,8 +81,8 @@ public:
 #endif
     virtual void deserialize(D4StreamUnMarshaller &um, DMR &dmr);
 
-    virtual unsigned int val2buf(void *val, bool reuse = false);
-    virtual unsigned int buf2val(void **val);
+    virtual uint64_t val2buf(void *val, bool reuse = false);
+    virtual uint64_t buf2val(void **val);
 
     virtual bool set_value(const dods_opaque &value);
     virtual dods_opaque value() const;
