+++ UnMarshaller.h
@@ -71,10 +71,10 @@ public:
     virtual void		get_opaque( char *val, unsigned int len ) = 0 ;
     virtual void		get_int( int &val ) = 0 ;
 
-    virtual void		get_vector( char **val, unsigned int &num,
+    virtual void		get_vector( char **val, uint64_t &num,
                                             Vector &vec ) = 0 ;
-    virtual void		get_vector( char **val, unsigned int &num,
-					    int width, Vector &vec ) = 0 ;
+    virtual void		get_vector( char **val, uint64_t &num,
+					    uint64_t width, Vector &vec ) = 0 ;
 
     virtual void		dump(std::ostream &strm) const = 0 ;
 } ;
